moved math helpers to helpers.ts
