+++ denograd/codegen/kernel.ts
@@ -1,7 +1,8 @@
 import { Device } from '../device.ts'
 import { ImageDType } from '../dtype.ts'
 import { all_int, all_same, AMX, ansilen, assert, cache, CAPTURE_PROCESS_REPLAY, colored, DEBUG, dedup, Enum, get_env, get_key, isinstance, isInt, NotImplemented, product, range, round_up, set_default, TC_OPT, to_function_name, USE_TC, WeakValueMap, zip } from '../helpers.ts'
-import { can_pad, graph_rewrite, GroupOp, idiv, KernelInfo, le, mod, mul, ne, Ops, print_uops, prod, resolve, type sint, UOp, type Variable, view_left } from '../ops.ts'
+import { can_pad, graph_rewrite, GroupOp, KernelInfo, Ops, print_uops, resolve, type sint, UOp, type Variable, view_left } from '../ops.ts'
+import { idiv, le, mod, mul, ne, prod } from '../helpers.ts'
 import { ProgramSpec, type Renderer, type TensorCore } from '../renderer/index.ts'
 import { ShapeTracker } from '../shape/shapetracker.ts'
 import { strides_for_shape } from '../shape/view.ts'
