+++ denograd/codegen/lowerer.ts
@@ -1,6 +1,6 @@
 import { dtypes, type PtrDType } from '../dtype.ts'
-import { all_int, is_eq, isinstance, min, partition, range, zip } from '../helpers.ts'
-import { graph_rewrite, identity_element, KernelInfo, Ops, PatternMatcher, prod, type sint, sint_to_uop, UOp, UPat } from '../ops.ts'
+import { all_int, is_eq, isinstance, min, partition, prod, range, zip } from '../helpers.ts'
+import { graph_rewrite, identity_element, KernelInfo, Ops, PatternMatcher, type sint, sint_to_uop, UOp, UPat } from '../ops.ts'
 import type { Renderer } from '../renderer/index.ts'
 
 // # returns the axes to create new_shape if new_shape can be created by combining axis from old_shape
