+++ denograd/engine/search.ts
@@ -1,7 +1,7 @@
 import type { Kernel } from '../codegen/kernel.ts'
 import { Buffer, type Compiler, Device, type Program } from '../device.ts'
-import { get_env, isInf, NotImplemented, range, zip } from '../helpers.ts'
-import { idiv, prod, sym_infer, type UOp, type Variable } from '../ops.ts'
+import { get_env, idiv, isInf, NotImplemented, prod, range, zip } from '../helpers.ts'
+import { sym_infer, type UOp, type Variable } from '../ops.ts'
 import type { ProgramSpec } from '../renderer/index.ts'
 import { Tensor } from '../tensor.ts'
 import { CompiledRunner } from './realize.ts'
