+++ denograd/nn/index.ts
@@ -1,7 +1,8 @@
 import { is_dtype_supported } from '../device.ts'
 import { dtypes } from '../dtype.ts'
 import { flatten, is_eq, make_tuple, range, zip } from '../helpers.ts'
-import { div, idiv, mul, prod, type sint, sub } from '../ops.ts'
+import { div, idiv, mul, prod, sub } from '../helpers.ts'
+import type { sint } from '../ops.ts'
 import { type Layer, Tensor } from '../tensor.ts'
 import { get_state_dict } from './state.ts'
 import { load_state_dict, safe_load, safe_save } from './state.ts'
