+++ denograd/ops.ts
@@ -1,7 +1,7 @@
 import { Buffer, type DeviceType } from './device.ts'
 import { type ConstType, DType, dtypes, ImageDType, PtrDType, truncate } from './dtype.ts'
 import { Env } from './env/index.ts'
-import { flatten, get_env, NotImplemented } from './helpers.ts'
+import { add, and, div, flatten, get_env, idiv, lshift, lt, mod, mul, ne, neg, NotImplemented, or, prod, rshift, sub, xor } from './helpers.ts'
 import { _METADATA, abs, all_int, all_same, assert, cache, CONSTS, counter, DEBUG, divmod, Enum, get_key, get_number_env, is_eq, is_less_than, is_subset, isInf, list_str, math_gcd, max, type Metadata, min, partition, permutations, range, set_default, sin, SPLIT_REDUCEOP, sqrt, trunc, WeakValueMap, zip } from './helpers.ts'
 import { ShapeTracker } from './shape/shapetracker.ts'
 
@@ -1429,42 +1429,3 @@ export const view_left = merge_views.add(
     [new UPat(GroupOp.Buffer).named('b').view(undefined, { name: 'v' }), ({ b, v }) => b.replace({ src: b.src.map((s) => s.op === Ops.VIEW ? (s.st!.add(v.st!)).to_uop() : s) })],
   ]),
 )
-
-export type Math = number | bigint | MathTrait<any>
-type Return<A, B> = A extends MathTrait<any> ? A : B extends MathTrait<any> ? B : A extends bigint ? bigint : B extends bigint ? bigint : number
-const _meta = (mathFn: (a: MathTrait<MathTrait<any>>, b: Math, reverse: boolean) => MathTrait<any>, numberFn: (a: number, b: number) => number, bigintFn?: (a: bigint, b: bigint) => bigint) => {
-  if (!bigintFn) bigintFn = numberFn as unknown as (a: bigint, b: bigint) => bigint
-  return <A extends Math, B extends Math>(a: A, b: B): Return<A, B> => {
-    if (a instanceof MathTrait) return mathFn(a, b, false) as Return<A, B>
-    else if (b instanceof MathTrait) return mathFn(b, a, true) as Return<A, B>
-    else if (typeof a === 'bigint' || typeof b === 'bigint') return bigintFn(BigInt(a), BigInt(b)) as Return<A, B>
-    else return numberFn(a as any, b as any) as Return<A, B>
-  }
-}
-
-export const add = _meta((a, b, r) => a.add(b, r), (a, b) => a + b)
-export const sub = _meta((a, b, r) => a.sub(b, r), (a, b) => a - b)
-export const mul = _meta((a, b, r) => a.mul(b, r), (a, b) => a * b)
-export const div = _meta((a, b, r) => a.div(b, r), (a, b) => a / b)
-export const idiv = _meta((a, b, r) => a.idiv(b, r), (a, b) => Math.floor(a / b), (a, b) => a / b)
-export const neg = <A extends Math>(a: A): Return<A, A> => ((typeof a !== 'number' && typeof a !== 'bigint') ? a.neg() : typeof a === 'bigint' ? a * -1n : a * -1)
-export const mod = _meta((a, b, r) => a.mod(b, r), (a, b) => a % b)
-
-export const and = _meta((a, b, r) => a.bitwise_and(b, r), (a, b) => a & b)
-export const or = _meta((a, b, r) => a.bitwise_or(b, r), (a, b) => a | b)
-export const xor = _meta((a, b, r) => a.xor(b, r), (a, b) => a ^ b)
-export const lshift = _meta((a, b, r) => a.lshift(b, r), (a, b) => a << b)
-export const rshift = _meta((a, b, r) => a.rshift(b, r), (a, b) => a >> b)
-
-export const lt = _meta((a, b, r) => !r ? a.lt(b) : a.const_like(b as any).lt(a), (a, b) => Number(a < b), (a, b) => BigInt(a < b))
-export const gt = _meta((a, b, r) => !r ? a.gt(b) : a.const_like(b as any).gt(a), (a, b) => Number(a > b), (a, b) => BigInt(a > b))
-export const le = _meta((a, b, r) => !r ? a.le(b) : a.const_like(b as any).le(a), (a, b) => Number(a <= b), (a, b) => BigInt(a <= b))
-export const ge = _meta((a, b, r) => !r ? a.ge(b) : a.const_like(b as any).ge(a), (a, b) => Number(a >= b), (a, b) => BigInt(a >= b))
-export const ne = _meta((a, b, r) => !r ? a.ne(b) : a.const_like(b as any).ne(a), (a, b) => Number(a !== b), (a, b) => BigInt(a !== b))
-export const eq = _meta((a, b, r) => !r ? a.eq(b) : a.const_like(b as any).eq(a), (a, b) => Number(a === b), (a, b) => BigInt(a === b))
-
-export const polyN = <T extends Math>(x: T, p: Math[]): T => p.reduce((acc, c) => add(mul(acc, x), c), 0) as T
-export const prod = <T extends Math>(x: T[]): T => x.reduce((acc, curr) => mul(acc, curr) as T, 1 as T)
-export const sum = <T extends Math>(x: T[]): T => x.reduce((acc, curr) => add(acc, curr) as T, 0 as T)
-export const ceildiv = <A extends Math, B extends Math>(num: A, amt: B): Return<A, B> => neg(idiv(num, neg(amt))) as Return<A, B>
-export const pow = _meta((a, b, r) => (a as any).pow(b, r), (a, b) => a ** b)
