+++ denograd/renderer/index.ts
@@ -1,7 +1,7 @@
 import type { DeviceType } from '../device.ts'
 import type { DType } from '../dtype.ts'
-import { assert, cache, dedup, get_key, range, replace, WeakValueMap } from '../helpers.ts'
-import { add, GroupOp, idiv, mul, Ops, prod, type sint, ssimplify, sym_infer, type UOp, type Variable } from '../ops.ts'
+import { add, assert, cache, dedup, get_key, idiv, mul, prod, range, replace, WeakValueMap } from '../helpers.ts'
+import { GroupOp, Ops, type sint, ssimplify, sym_infer, type UOp, type Variable } from '../ops.ts'
 
 export type TC = [number, number]
